一些错误的修复之类的

+++ src/log.rs
@@ -1,67 +1,70 @@
 use std::{fs::File, io::Write, path::Path};
 
 use chrono::{DateTime, Local};
+use env_logger::Builder;
+use log::info;
 use log4rs::{
-    append::file::FileAppender,
+    append::{file::FileAppender, console::ConsoleAppender},
     config::{Appender, Root},
     encode::pattern::PatternEncoder,
     Config,
 };
 
-pub fn init() {
+use crate::error::KappaError;
+
+pub fn init() -> Result<(), Box<dyn std::error::Error>> {
+    // 配置日志系统
     let log_path = Path::new("logs/latest.log");
+    // 将上次的日志保存至新文件
     if log_path.exists() {
+        // 获取老日志访问时间，用来保存老日志
         let time: DateTime<Local> = log_path
             .metadata()
             .expect("Could not read metatata of last log file!")
-            .modified()
-            .unwrap()
+            .modified()?
             .into();
-        let time = time.format("%Y-%m-%dT%H:%M:%S");
+        let time = time.format("%Y%m%d-%H%M%S");
 
         let mut i = 0;
         let mut last_file = File::create(format!("logs/{}.log", time));
-        while last_file.is_err() || i < 16 {
+        while last_file.is_err() && i < 16 {
             i += 1;
             last_file = File::create(format!("logs/{}-{}.log", time, i));
         }
 
-        let mut last_file = last_file.unwrap();
+        if i >= 15 {
+            return Err(Box::new(KappaError::new("Could not save last log file!")));
+        }
+
+        let mut last_file = last_file?;
         last_file
-            .write(&std::fs::read(log_path).unwrap())
+            .write(&std::fs::read(log_path)?)
             .expect("Could not copy to log file! ");
         std::fs::remove_file(Path::new("logs/latest.log")).expect("Could not remove old log file!");
     }
 
-    let log_file = FileAppender::builder()
+    let file_appender = FileAppender::builder()
         .encoder(Box::new(PatternEncoder::new(
             "[{d(%Y-%m-%dT%H:%M:%S)}] [{l}] [{t}]: {m}{n}",
         )))
-        .build("logs/latest.log")
-        .unwrap();
+        .build("logs/latest.log")?;
+    
+    let console_appender = ConsoleAppender::builder()
+        .encoder(Box::new(PatternEncoder::new(
+            "[{d(%Y-%m-%dT%H:%M:%S)}] [{l}] [{t}]: {m}{n}",
+        )))
+        .build();
 
     let config = Config::builder()
-        .appender(Appender::builder().build("logfile", Box::new(log_file)))
+        .appenders(vec![Appender::builder().build("console", Box::new(console_appender)), Appender::builder().build("logfile", Box::new(file_appender))])
         .build(
             Root::builder()
-                .appender("logfile")
+                .appenders(vec!["logfile", "console"])
                 .build(log::LevelFilter::Info),
-        )
-        .unwrap();
+        )?;
 
-    log4rs::init_config(config).unwrap();
-
-    // Builder::new()
-    //     .format(|buf, record| {
-    //         writeln!(
-    //             buf,
-    //             "[{}][{}]: {}",
-    //             chrono::Local::now().format("%Y-%m-%dT%H:%M:%S"),
-    //             record.level(),
-    //             record.args(),
-    //         )
-    //     })
-    //     .filter(None, log::LevelFilter::Info)
-    //     .format_timestamp(None)
-    //     .init();
+    log4rs::init_config(config)?;
+    
+    info!("Logging system initialized.");
+    Ok(())
 }
